restructure + nodenext compatibility
