+++ .github/workflows/ci.yml
@@ -14,4 +14,5 @@ jobs:
   test:
     uses: fastify/workflows/.github/workflows/plugins-ci.yml@v3
     with:
+      lint: true
       license-check: true
