+++ .taprc
@@ -1,0 +1,1 @@
+--no-coverage
