+++ types/index.d.ts
@@ -1,0 +1,31 @@
+import { Application } from 'express'
+import { FastifyPluginCallback } from 'fastify'
+
+declare module 'fastify' {
+  interface FastifyInstance {
+    /**
+     * Express middleware function
+     */
+    use: Application['use']
+
+    /**
+     * Express application instance
+     */
+    express: Application
+  }
+}
+
+type FastifyExpress = FastifyPluginCallback<fastifyExpress.FastifyExpressOptions>
+
+declare namespace fastifyExpress {
+  
+  export interface FastifyExpressOptions {
+    expressHook?: string;
+  }
+
+  export const fastifyExpress: FastifyExpress
+  export { fastifyExpress as default }
+}
+
+declare function fastifyExpress(...params: Parameters<FastifyExpress>): ReturnType<FastifyExpress>
+export = fastifyExpress
